+++ wallet/src/state.rs
@@ -35,7 +35,7 @@ const SUBMITTED_TRANSACTION_TIMEOUT: Duration = Duration::from_secs(60);
 )]
 pub struct ClientState {
     /// The last block height we've scanned to, if any.
-    last_block_height: Option<u32>,
+    last_block_height: Option<u64>,
     /// Note commitment tree.
     note_commitment_tree: NoteCommitmentTree,
     /// Our nullifiers and the notes they correspond to.
@@ -567,7 +567,7 @@ impl ClientState {
     }
 
     /// Returns the last block height the client state has synced up to, if any.
-    pub fn last_block_height(&self) -> Option<u32> {
+    pub fn last_block_height(&self) -> Option<u64> {
         self.last_block_height
     }
 
@@ -769,7 +769,7 @@ mod serde_helpers {
     #[serde_as]
     #[derive(Serialize, Deserialize)]
     pub struct ClientStateHelper {
-        last_block_height: Option<u32>,
+        last_block_height: Option<u64>,
         #[serde_as(as = "serde_with::hex::Hex")]
         note_commitment_tree: Vec<u8>,
         nullifier_map: Vec<(String, String)>,
